requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
supabase==2.0.3
schedule==1.2.1
gunicorn==21.2.0
